fix(offers): drop unnecessary Integer→long boxing in countOffers (PMD gate)
Latent develop-red: PMD 7.7.0 UnnecessaryBoxing flags totalCount.longValue();
the ternary against the long EMPTY_COUNT already unboxes-and-widens.

+++ allegro-client/src/main/java/io/github/mgrtomaszzurawski/allegro/sdk/internal/client/offers/OffersImpl.java
@@ -226,7 +226,7 @@ public final class OffersImpl implements Offers {
         Integer totalCount = response.getTotalCount();
         // The offers listing always carries totalCount (ADR-010 only wires this accessor where the
         // server reports one); the null branch is a defensive guard, not a supported "no total" signal.
-        return totalCount == null ? EMPTY_COUNT : totalCount.longValue();
+        return totalCount == null ? EMPTY_COUNT : totalCount;
     }
 
     private Query filterQuery(OfferFilter filter) {
